requirements.txt
django==3.1.2
pandas==1.1.5
uWSGI
djangorestframework
markdown>=3.0.0
django-filter>=1.1.1
